% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsize.R
\name{print.wsize}
\alias{print.wsize}
\title{Print Method for `wsize` Objects}
\usage{
\method{print}{wsize}(x, ...)
}
\arguments{
\item{x}{An object of class \code{wsize}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Prints the sample size determination results to the console.
}
\description{
Custom print method for objects of class `wsize`.
Displays the results of the sample size determination in a user-friendly format.
}
\examples{
\donttest{
# Print the results from the wsize function
n <- c(10, 10, 10, 10)
means <- c(1, 0, 0, -1)
sd <- c(1, 1, 1, 1)
result <- wanova_pwr.test(n, means, sd, power = 0.90, alpha = 0.05)
print(result)
}

}
